+++ src/main/java/org/jabref/gui/maintable/MainTable.java
@@ -12,6 +12,7 @@ import javax.swing.undo.UndoManager;
 
 import javafx.collections.ListChangeListener;
 import javafx.scene.control.SelectionMode;
+import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableRow;
 import javafx.scene.control.TableView;
 import javafx.scene.input.ClipboardContent;
@@ -57,10 +58,20 @@ public class MainTable extends TableView<BibEntryTableViewModel> {
     private final ImportHandler importHandler;
     private final CustomLocalDragboard localDragboard;
 
+    private long lastKeyPressTime;
+    private String columnSearchTerm;
+
     public MainTable(MainTableDataModel model, JabRefFrame frame,
                      BasePanel panel, BibDatabaseContext database,
                      MainTablePreferences preferences, ExternalFileTypes externalFileTypes, KeyBindingRepository keyBindingRepository) {
         super();
+
+        this.setOnKeyTyped(key -> {
+            if (this.getSortOrder().isEmpty()) {
+                return;
+            }
+            this.jumpToSearchKey(getSortOrder().get(0), key);
+        });
 
         this.model = model;
         this.database = Objects.requireNonNull(database);
@@ -121,6 +132,40 @@ public class MainTable extends TableView<BibEntryTableViewModel> {
         setupKeyBindings(keyBindingRepository);
 
         database.getDatabase().registerListener(this);
+    }
+
+    /**
+     * This is called, if a user starts typing some characters into the keyboard with focus on main table.
+     * The {@link MainTable} will scroll to the cell with the same starting column value and typed string
+     *
+     * @param sortedColumn The sorted column in {@link MainTable}
+     * @param keyEvent The pressed character
+     */
+
+    private void jumpToSearchKey(TableColumn<BibEntryTableViewModel, ?> sortedColumn, KeyEvent keyEvent) {
+        if (keyEvent.getCharacter() == null || sortedColumn == null) {
+            return;
+        }
+
+        if (System.currentTimeMillis() - lastKeyPressTime < 700) {
+            columnSearchTerm += keyEvent.getCharacter().toLowerCase();
+        } else {
+            columnSearchTerm = keyEvent.getCharacter().toLowerCase();
+        }
+
+        lastKeyPressTime = System.currentTimeMillis();
+
+        this.getItems().stream()
+            .filter(item -> Optional.ofNullable(sortedColumn.getCellObservableValue(item).getValue())
+                                    .map(Object::toString)
+                                    .orElse("")
+                                    .toLowerCase()
+                                    .startsWith(columnSearchTerm))
+            .findFirst()
+            .ifPresent(item -> {
+                this.scrollTo(item);
+                this.clearAndSelect(item.getEntry());
+            });
     }
 
     @Subscribe